Refactoring array filter API
The API for the query operator used when a filter returns an array has
been narrowed to only :all and :any, where :all maps to the MongoDB $all
operator and :any maps to the MongoDB $in operator. Note that when no
operator is passed, the default is $all, which preserves backwards
compatability. Also, when an unknown operator is passed, an error is
raised.